Perf: Allow preproc defines in the shaders to section of areas of the code that won't be used in certain situations. Didn't give me as much of a win as I wanted though.

+++ digitanks/src/digitanks/digitanksgame.cpp
@@ -974,7 +974,7 @@ void CDigitanksGame::SetupMenuMarch()
 				pMarcher = GameServer()->Create<CMenuMarcher>("CMenuMarcher");
 				pTeam->AddEntity(pMarcher);
 
-				pMarcher->SetGlobalOrigin(GetTerrain()->GetPointHeight(Vector(RemapVal((float)j, 0, iColumns, -flSpread, flSpread), flY + RemapVal((float)k, 0, iRows, -15, 15), 0)));
+				pMarcher->SetGlobalOrigin(GetTerrain()->GetPointHeight(Vector(RemapVal((float)j, 0, (float)iColumns, -flSpread, flSpread), flY + RemapVal((float)k, 0, (float)iRows, -15, 15), 0)));
 				pMarcher->SetGlobalAngles(EAngle(0,90,0));
 
 				if (GameServer()->GetWorkListener())
